atualização app

+++ app.py
@@ -81,6 +81,9 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
+
+    if id_user == current_user.id:
+        return jsonify({"message": "Não é permitido excluir o usuário logado!"})
     if user:
         db.session.delete(user)
         db.session.commit()
